Introduce new OpenRA.exe entrypoint and change OpenRA.Game to a library.
