+++ OpenRA.Launcher/App.config
