+++ OpenRA.Launcher/OpenRA.Launcher.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <OutputType>Library</OutputType>
+    <OutputType>Exe</OutputType>
     <TargetFramework>net472</TargetFramework>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <Optimize>true</Optimize>
@@ -9,7 +9,6 @@
     <DebugSymbols>true</DebugSymbols>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <AutoGenerateBindingRedirects>false</AutoGenerateBindingRedirects>
-    <RootNamespace>OpenRA</RootNamespace>
     <OutputPath>../bin</OutputPath>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
     <PlatformTarget>AnyCPU</PlatformTarget>
@@ -17,6 +16,7 @@
     <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
     <CodeAnalysisRuleSet>..\OpenRA.ruleset</CodeAnalysisRuleSet>
     <Configurations>Release;Debug</Configurations>
+    <AssemblyName>OpenRA</AssemblyName>
   </PropertyGroup>
   <PropertyGroup>
     <TargetPlatform Condition="$([MSBuild]::IsOsPlatform('Windows'))">win-x64</TargetPlatform>
@@ -31,13 +31,30 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <Optimize>false</Optimize>
   </PropertyGroup>
+  <PropertyGroup Condition="'$(TargetPlatform)' == 'win-x86'">
+    <Prefer32bit>true</Prefer32bit>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(RunConfiguration)' == 'Red Alert'">
+    <StartAction>Project</StartAction>
+    <StartArguments>Engine.EngineDir=".." Game.Mod=ra</StartArguments>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(RunConfiguration)' == 'Dune 2000'">
+    <StartAction>Project</StartAction>
+    <StartArguments>Engine.EngineDir=".." Game.Mod=d2k</StartArguments>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(RunConfiguration)' == 'Tiberian Dawn'">
+    <StartAction>Project</StartAction>
+    <StartArguments>Engine.EngineDir=".." Game.Mod=cnc</StartArguments>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(RunConfiguration)' == 'Tiberian Sun'">
+    <StartAction>Project</StartAction>
+    <StartArguments>Engine.EngineDir=".." Game.Mod=ts</StartArguments>
+  </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="OpenRA-Eluant" Version="1.0.17" />
-    <PackageReference Include="OpenRA-Open.NAT" Version="1.0.0" />
-    <PackageReference Include="SharpZipLib" Version="1.2.0" />
-    <PackageReference Include="StyleCop.Analyzers" Version="1.1.118" />
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0" PrivateAssets="All" />
+    <None Include="App.config" />
+    <ProjectReference Include="..\OpenRA.Game\OpenRA.Game.csproj" />
     <AdditionalFiles Include="../stylecop.json" />
+    <AdditionalFiles Include="Properties/launchSettings.json" />
   </ItemGroup>
   <Target Name="DisableAnalyzers" BeforeTargets="CoreCompile" Condition="'$(Configuration)'=='Release'">
     <!-- Disable code style analysis on Release builds to improve compile-time performance -->
